+++ src/app/common/directives/form-validation-errors.types.ts
@@ -15,7 +15,10 @@ export type ValidationError =
     | "email"
     | "min"
     | "max"
-    | "whitespace";
+    | "whitespace"
+    | "range"
+    | "invalidCronExpression"
+    | "noneOf";
 
 export type ValidationErrorTuple = {
     error: ValidationError;